+++ CHANGELOG.md
@@ -233,7 +233,8 @@ Users/clients that do not make use of these endpoints will not be impacted.
 
 - **CUMULUS-3223**
   - Update `@cumulus/cmrjs/cmr-utils.getGranuleTemporalInfo` to handle the error when the cmr file s3url is not available
-  - Update `sfEventSqsToDbRecords` lambda to return partial batch failure
+  - Update `sfEventSqsToDbRecords` lambda to return [partial batch failure](https://docs.aws.amazon.com/lambda/latest/dg/with-sqs.html#services-sqs-batchfailurereporting),
+    and only reprocess messages when cumulus message can't be retrieved from the execution events.
   - Update `@cumulus/cumulus-message-adapter-js` to `2.0.5` for all cumulus tasks
 
 ### Removed